+++ src/Type/Collection/Exception/NumericCollectionFactoryException.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Type\Collection\Exception;
+
+use LDL\Framework\Base\Exception\LDLFactoryException;
+
+class NumericCollectionFactoryException extends LDLFactoryException
+{
+
+}