+++ src/Type/Collection/Interfaces/Factory/UniqueStringCollectionFactoryInterface.php
@@ -1,0 +1,17 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Type\Collection\Interfaces\Factory;
+
+use LDL\Framework\Base\Contracts\FactoryInterface;
+use LDL\Type\Collection\Types\String\UniqueStringCollection;
+use LDL\Type\Collection\Exception\UniqueStringCollectionFactoryException;
+
+interface UniqueStringCollectionFactoryInterface extends FactoryInterface
+{    
+    /**
+     * @param  UniqueStringCollection $collection
+     * @throws UniqueStringCollectionFactoryException
+     * @return mixed
+     */
+    public static function fromUniqueStringCollection(UniqueStringCollection $collection);
+}